CompleteARR V1.0 - First Public Release
This is the first public upload of the CompleteARR Suite.
This was created to run on Win10/11
Please read the included docs. I tried to include everything a new user would need to know. 

Having issues or have coding advice to make any of the scripts work better? Please message me. 
This is my first project of this kind, help is appreciated and welcome. Just be nice about it. 

Future Release Plans
- Docker Container Version
- User Interface for users who don't like editing YML files